move modal sources: vm-export
